+++ src/Serilog/Parsing/MessageTemplateParser.cs
@@ -52,8 +52,13 @@ namespace Serilog.Parsing
 #if FEATURE_SPAN
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         static MessageTemplateToken[] Tokenize(string messageTemplate) => Tokenize(messageTemplate.AsSpan());
+#endif
 
+#if FEATURE_SPAN
         static MessageTemplateToken[] Tokenize(ReadOnlySpan<char> messageTemplate)
+#else
+        static MessageTemplateToken[] Tokenize(string messageTemplate)
+#endif
         {
             var tokens = new List<MessageTemplateToken>();
             var reusableAccumInstance = new StringBuilder();
@@ -80,7 +85,11 @@ namespace Serilog.Parsing
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#if FEATURE_SPAN
         static MessageTemplateToken ParsePropertyToken(int startAt, ReadOnlySpan<char> messageTemplate, out int next)
+#else
+        static MessageTemplateToken ParsePropertyToken(int startAt, string messageTemplate, out int next)
+#endif
         {
             var first = startAt;
             startAt++;
@@ -155,14 +164,18 @@ namespace Serilog.Parsing
             return new PropertyToken(
                 propertyName.ToString(),
                 rawText.ToString(),
-                format.IsEmpty ? null : format.ToString(),
+                format.IsEmpty() ? null : format.ToString(),
                 alignmentValue,
                 destructuring,
                 first);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#if FEATURE_SPAN
         static bool TrySplitTagContent(ReadOnlySpan<char> tagContent, out ReadOnlySpan<char> propertyNameAndDestructuring, out ReadOnlySpan<char> format, out ReadOnlySpan<char> alignment)
+#else
+        static bool TrySplitTagContent(string tagContent, out string propertyNameAndDestructuring, out string format, out string alignment)
+#endif
         {
             var formatDelim = tagContent.IndexOf(':');
             var alignmentDelim = tagContent.IndexOf(',');
@@ -216,217 +229,11 @@ namespace Serilog.Parsing
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#if FEATURE_SPAN
         static TextToken ParseTextToken(int startAt, ReadOnlySpan<char> messageTemplate, out int next, StringBuilder reusableAccumInstance)
-        {
-            var first = startAt;
-
-            reusableAccumInstance.Clear();
-            var accum = reusableAccumInstance;
-
-            do
-            {
-                var nc = messageTemplate[startAt];
-                if (nc == '{')
-                {
-                    if (startAt + 1 < messageTemplate.Length &&
-                        messageTemplate[startAt + 1] == '{')
-                    {
-                        accum.Append(nc);
-                        startAt++;
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-                else
-                {
-                    accum.Append(nc);
-                    if (nc == '}')
-                    {
-                        if (startAt + 1 < messageTemplate.Length &&
-                            messageTemplate[startAt + 1] == '}')
-                        {
-                            startAt++;
-                        }
-                    }
-                }
-
-                startAt++;
-            } while (startAt < messageTemplate.Length);
-
-            next = startAt;
-
-            if (first == next)
-                return null;
-
-            return new TextToken(accum.ToString(), first);
-        }
 #else
-        static MessageTemplateToken[] Tokenize(string messageTemplate)
-        {
-            var tokens = new List<MessageTemplateToken>();
-            var reusableAccumInstance = new StringBuilder();
-
-            var nextIndex = 0;
-            while (true)
-            {
-                var beforeText = nextIndex;
-                var tt = ParseTextToken(nextIndex, messageTemplate, out nextIndex, reusableAccumInstance);
-                if (nextIndex > beforeText)
-                    tokens.Add(tt);
-
-                if (nextIndex == messageTemplate.Length)
-                    return tokens.ToArray();
-
-                var beforeProp = nextIndex;
-                var pt = ParsePropertyToken(nextIndex, messageTemplate, out nextIndex);
-                if (beforeProp < nextIndex)
-                    tokens.Add(pt);
-
-                if (nextIndex == messageTemplate.Length)
-                    return tokens.ToArray();
-            }
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static MessageTemplateToken ParsePropertyToken(int startAt, string messageTemplate, out int next)
-        {
-            var first = startAt;
-            startAt++;
-            while (startAt < messageTemplate.Length && IsValidInPropertyTag(messageTemplate[startAt]))
-                startAt++;
-
-            if (startAt == messageTemplate.Length || messageTemplate[startAt] != '}')
-            {
-                next = startAt;
-                return new TextToken(messageTemplate.Substring(first, next - first), first);
-            }
-
-            next = startAt + 1;
-
-            var rawText = messageTemplate.Substring(first, next - first);
-            var tagContent = rawText.Substring(1, next - (first + 2));
-            if (tagContent.Length == 0)
-                return new TextToken(rawText, first);
-
-            if (!TrySplitTagContent(tagContent, out var propertyNameAndDestructuring, out var format, out var alignment))
-                return new TextToken(rawText, first);
-
-            var propertyName = propertyNameAndDestructuring;
-            var destructuring = Destructuring.Default;
-            if (propertyName.Length != 0 && TryGetDestructuringHint(propertyName[0], out destructuring))
-                propertyName = propertyName.Substring(1);
-
-            if (propertyName.Length == 0)
-                return new TextToken(rawText, first);
-
-            for (var i = 0; i < propertyName.Length; ++i)
-            {
-                var c = propertyName[i];
-                if (!IsValidInPropertyName(c))
-                    return new TextToken(rawText, first);
-            }
-
-            if (format != null)
-            {
-                for (var i = 0; i < format.Length; ++i)
-                {
-                    var c = format[i];
-                    if (!IsValidInFormat(c))
-                        return new TextToken(rawText, first);
-                }
-            }
-
-            Alignment? alignmentValue = null;
-            if (alignment != null)
-            {
-                for (var i = 0; i < alignment.Length; ++i)
-                {
-                    var c = alignment[i];
-                    if (!IsValidInAlignment(c))
-                        return new TextToken(rawText, first);
-                }
-
-                var lastDash = alignment.LastIndexOf('-');
-                if (lastDash > 0)
-                    return new TextToken(rawText, first);
-
-                if (!IntHelper.TryParse(lastDash == -1 ? alignment : alignment.Substring(1), out var width) || width == 0)
-                    return new TextToken(rawText, first);
-
-                var direction = lastDash == -1 ?
-                    AlignmentDirection.Right :
-                    AlignmentDirection.Left;
-
-                alignmentValue = new Alignment(direction, width);
-            }
-
-            return new PropertyToken(
-                propertyName,
-                rawText,
-                format,
-                alignmentValue,
-                destructuring,
-                first);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static bool TrySplitTagContent(string tagContent, out string propertyNameAndDestructuring, out string format, out string alignment)
-        {
-            var formatDelim = tagContent.IndexOf(':');
-            var alignmentDelim = tagContent.IndexOf(',');
-            if (formatDelim == -1 && alignmentDelim == -1)
-            {
-                propertyNameAndDestructuring = tagContent;
-                format = null;
-                alignment = null;
-            }
-            else
-            {
-                if (alignmentDelim == -1 || (formatDelim != -1 && alignmentDelim > formatDelim))
-                {
-                    propertyNameAndDestructuring = tagContent.Substring(0, formatDelim);
-                    format = formatDelim == tagContent.Length - 1 ?
-                        null :
-                        tagContent.Substring(formatDelim + 1);
-                    alignment = null;
-                }
-                else
-                {
-                    propertyNameAndDestructuring = tagContent.Substring(0, alignmentDelim);
-                    if (formatDelim == -1)
-                    {
-                        if (alignmentDelim == tagContent.Length - 1)
-                        {
-                            alignment = format = null;
-                            return false;
-                        }
-
-                        format = null;
-                        alignment = tagContent.Substring(alignmentDelim + 1);
-                    }
-                    else
-                    {
-                        if (alignmentDelim == formatDelim - 1)
-                        {
-                            alignment = format = null;
-                            return false;
-                        }
-
-                        alignment = tagContent.Substring(alignmentDelim + 1, formatDelim - alignmentDelim - 1);
-                        format = formatDelim == tagContent.Length - 1 ?
-                            null :
-                            tagContent.Substring(formatDelim + 1);
-                    }
-                }
-            }
-
-            return true;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         static TextToken ParseTextToken(int startAt, string messageTemplate, out int next, StringBuilder reusableAccumInstance)
+#endif
         {
             var first = startAt;
 
@@ -472,8 +279,6 @@ namespace Serilog.Parsing
 
             return new TextToken(accum.ToString(), first);
         }
-
-#endif
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         static bool TryGetDestructuringHint(char c, out Destructuring destructuring)
@@ -512,6 +317,6 @@ namespace Serilog.Parsing
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         static bool IsValidInFormat(char c) => c != '}' && (c == ' ' || c == '+' || char.IsLetterOrDigit(c) || char.IsPunctuation(c));
-
+        
     }
 }
