--- src/Serilog/Suport/IntHelper.cs
@@ -1,27 +1,0 @@
-using System;
-using System.Globalization;
-using System.Runtime.CompilerServices;
-
-namespace Serilog.Support
-{
-    static class IntHelper
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool TryParse(string str, out int result)
-        {
-            return int.TryParse(str, NumberStyles.None, CultureInfo.InvariantCulture, out result);
-        }
-
-#if FEATURE_SPAN
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool TryParse(ReadOnlySpan<char> str, out int result)
-        {
-#if FEATURE_SPAN_COMPLETE
-            return int.TryParse(str, NumberStyles.None, CultureInfo.InvariantCulture, out result);
-#else
-            return int.TryParse(str.ToString(), NumberStyles.None, CultureInfo.InvariantCulture, out result);
-#endif
-        }
-#endif
-    }
-}
